+++ services/app-api/package.json
@@ -37,6 +37,7 @@
         "eta": "^1.12.3",
         "graphql": "^16.2.0",
         "graphql-scalars": "^1.11.1",
+        "json2csv": "^5.0.7",
         "launchdarkly-node-server-sdk": "6.4.2",
         "neverthrow": "^4.2.2",
         "protobufjs": "^6.11.2",
@@ -50,6 +51,7 @@
         "@types/aws-lambda": "^8.10.83",
         "@types/glob": "^7.2.0",
         "@types/jest": "^27.0.1",
+        "@types/json2csv": "^5.0.3",
         "@types/request": "^2.48.7",
         "@types/uuid": "^8.3.1",
         "@typescript-eslint/eslint-plugin": "^5.6.0",
